refactor code stage 1
